changes output of convertCrayonValues func to CGFloat. Creates func for setInitialDetails in DetailsVC

+++ Pursuit-Core-iOS-Unit2Final/Pursuit-Core-iOS-Unit2Final/Controllers/ViewController.swift
@@ -49,7 +49,7 @@ extension CrayonListViewController: UITableViewDataSource, UITableViewDelegate {
         
         cell.textLabel?.text = crayon.name
         cell.detailTextLabel?.text = crayon.hex
-        cell.backgroundColor = UIColor(displayP3Red: CGFloat(redValue), green: CGFloat(greenValue), blue: CGFloat(blueValue), alpha: alphaValue)
+        cell.backgroundColor = UIColor(displayP3Red: redValue, green: greenValue, blue: blueValue, alpha: alphaValue)
         
         return cell
     }
